requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt
psycopg2